refactor to use runReadlineFinal
See commits leading up to 0.2.0.0 release of polysemy-readline for more
context. The short summary is that runReadline was resulting in
runInputT being run for each primitive effect. This was causing
performance overhead and also resulted in somethings not getting saved
properly (e.g. typing multiple commands while one is still executing).
All of these things appear to be fixed now.