switch to strict ByteString in most places
Except when generating bytestrings purely, it is ideal for bytestrings
to be read strictly to avoid lazy io. Switch most uses of LByteString
out for ByteString to enforce this. However, on output side stuff, it is
more okay to use LByteString, because it is essentially free to convert
strict ByteString into LByteString, while it might actually be more
efficient to construct a lazy bytestring (i.e. by a builder). So in some
places I have left it when it was required for a function outputing a
string.