Add OsPath helpers for future transition
I think it might not be viable to migrate until I can switch to
GHC-9.10+. Currently `System.OsString` is missing some functionality I
need to be able to convert between text/OsPath seamlessly. It isn't
possible to pin the packages because os-string + filesystem are boot
libraries in future GHC versions.